:hammer: Add URL generator to helper folder
